+++ Bots/DervFeatherFarm.py
@@ -1,215 +1,21 @@
 from Py4GWCoreLib import *
 
-# from Bots.DervFeatherFarmer import DervFeatherFarmer
-# from Bots.DervFeatherFarmer import DervBuildFarmStatus
-
-from Py4GWCoreLib import GLOBAL_CACHE
-from Py4GWCoreLib import Routines
-from Py4GWCoreLib import BuildMgr
-from Py4GWCoreLib import ActionQueueManager
-from Py4GWCoreLib import Range
-from Py4GWCoreLib import Profession
-from Py4GWCoreLib import Keystroke
-from Py4GWCoreLib import Key
-from Py4GWCoreLib import Player
-from Py4GWCoreLib import Weapon
-from Py4GWCoreLib.Builds.AutoCombat import AutoCombat
-
-
-# =================== BUILD ========================
-class DervBuildFarmStatus:
-    Move = 'move'
-    Ball = 'ball'
-    Kill = 'kill'
-
-
-class DervFeatherFarmer(BuildMgr):
-    def __init__(self):
-        super().__init__(
-            name="Derv Feather Farmer",
-            required_primary=Profession.Dervish,
-            required_secondary=Profession.Assassin,
-            template_code='OgejkmrMbSmXfbaXNXTQ3lEYsXA',
-            skills=[
-                GLOBAL_CACHE.Skill.GetID("Sand_Shards"),
-                GLOBAL_CACHE.Skill.GetID("Vow_of_Strength"),
-                GLOBAL_CACHE.Skill.GetID("Staggering_Force"),
-                GLOBAL_CACHE.Skill.GetID("Eremites_Attack"),
-                GLOBAL_CACHE.Skill.GetID("Dash"),
-                GLOBAL_CACHE.Skill.GetID("Dwarven_Stability"),
-                GLOBAL_CACHE.Skill.GetID("Conviction"),
-                GLOBAL_CACHE.Skill.GetID("Mystic_Regeneration"),
-            ],
-        )
-
-        self.auto_combat_handler: BuildMgr = AutoCombat()
-        # assign extra skill attributes from the already populated self.skills
-        self.sand_shards = self.skills[0]
-        self.vow_of_strength = self.skills[1]
-        self.staggering_force = self.skills[2]
-        self.eremites_attack = self.skills[3]
-        self.dash = self.skills[4]
-        self.dwarven_stability = self.skills[5]
-        self.conviction = self.skills[6]
-        self.mystic_regen = self.skills[7]
-
-        # Build usage status
-        self.status = DervBuildFarmStatus.Move
-        self.spiked = False
-        self.spiking = False
-
-    def swap_to_scythe(self):
-        if GLOBAL_CACHE.Agent.GetWeaponType(Player.GetAgentID())[0] != Weapon.Scythe:
-            Keystroke.PressAndRelease(Key.F1.value)
-            yield
-
-    def swap_to_shield_set(self):
-        if GLOBAL_CACHE.Agent.GetWeaponType(Player.GetAgentID())[0] == Weapon.Scythe:
-            Keystroke.PressAndRelease(Key.F2.value)
-            self.status = DervBuildFarmStatus.Move
-            yield from Routines.Yield.wait(750)
-
-    def is_target_sensali(self, agent_id):
-        if not agent_id:
-            return False
-
-        SENSALI = 'Sensali'
-        if SENSALI in GLOBAL_CACHE.Agent.GetName(agent_id):
-            return True
-        return False
-
-    def get_sensali_target(self, agent_ids):
-        for agent_id in agent_ids:
-            if self.is_target_sensali(agent_id):
-                return agent_id
-        return None
-
-    def ProcessSkillCasting(self):
-        if not (
-            Routines.Checks.Map.IsExplorable()
-            and Routines.Checks.Player.CanAct()
-            and Routines.Checks.Map.IsExplorable()
-            and Routines.Checks.Skills.CanCast()
-        ):
-            ActionQueueManager().ResetAllQueues()
-            yield from Routines.Yield.wait(1000)
-            return
-
-        player_agent_id = GLOBAL_CACHE.Player.GetAgentID()
-        has_dwarven_stability = Routines.Checks.Effects.HasBuff(player_agent_id, self.dwarven_stability)
-        has_mystic_regen = Routines.Checks.Effects.HasBuff(player_agent_id, self.mystic_regen)
-        has_conviction = Routines.Checks.Effects.HasBuff(player_agent_id, self.conviction)
-
-        if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.dwarven_stability)) and not has_dwarven_stability:
-            yield from Routines.Yield.Skills.CastSkillID(self.dwarven_stability, aftercast_delay=100)
-            return
-
-        player_hp = GLOBAL_CACHE.Agent.GetHealth(GLOBAL_CACHE.Player.GetAgentID())
-        if (
-            (yield from Routines.Yield.Skills.IsSkillIDUsable(self.mystic_regen))
-            and not has_mystic_regen
-            and player_hp < 0.95
-        ):
-            yield from Routines.Yield.Skills.CastSkillID(self.mystic_regen, aftercast_delay=750)
-            return
-
-        if self.status == DervBuildFarmStatus.Move:
-            yield from self.swap_to_shield_set()
-            self.spiked = False
-            if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.dash)) and has_dwarven_stability:
-                yield from Routines.Yield.Skills.CastSkillID(self.dash, aftercast_delay=100)
-                return
-
-        if self.status == DervBuildFarmStatus.Ball:
-            yield from self.swap_to_shield_set()
-            self.spiked = False
-            if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.conviction)):
-                yield from Routines.Yield.Skills.CastSkillID(self.conviction, aftercast_delay=750)
-                return
-
-        if self.status == DervBuildFarmStatus.Kill:
-            player_pos = GLOBAL_CACHE.Player.GetXY()
-            enemies = Routines.Agents.GetFilteredEnemyArray(player_pos[0], player_pos[1], Range.Earshot.value)
-            target_sensali = self.get_sensali_target(enemies)
-
-            if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.conviction)) and not has_conviction:
-                yield from Routines.Yield.Skills.CastSkillID(self.conviction, aftercast_delay=1000)
-                return
-
-            player_current_energy = GLOBAL_CACHE.Agent.GetEnergy(player_agent_id) * GLOBAL_CACHE.Agent.GetMaxEnergy(
-                player_agent_id
-            )
-            if self.spiking or (not self.spiked and target_sensali and player_current_energy >= 25):
-                print('Currently Spiking!')
-                self.spiking = True
-                has_sand_shards = Routines.Checks.Effects.HasBuff(player_agent_id, self.sand_shards)
-
-                if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.sand_shards)):
-                    print('Casting Sand Shards!')
-                    yield from Routines.Yield.Skills.CastSkillID(self.sand_shards, aftercast_delay=250)
-                    return
-
-                if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.vow_of_strength)) and has_sand_shards:
-                    print('Casting Vow of Strength!')
-                    yield from Routines.Yield.Skills.CastSkillID(self.vow_of_strength, aftercast_delay=750)
-                    return
-
-                has_vow_of_strength = Routines.Checks.Effects.HasBuff(player_agent_id, self.vow_of_strength)
-                if (
-                    (
-                        yield from Routines.Yield.Skills.IsSkillIDUsable(self.staggering_force)
-                        and Routines.Yield.Skills.IsSkillIDUsable(self.eremites_attack)
-                    )
-                    and has_vow_of_strength
-                    and has_sand_shards
-                ):
-                    yield from Routines.Yield.Agents.TargetNearestEnemy(Range.Earshot.value)
-                    print('Spiking with Staggering Force and Eremites!')
-                    yield from Routines.Yield.Skills.CastSkillID(self.staggering_force, aftercast_delay=250)
-                    has_staggering_force = Routines.Checks.Effects.HasBuff(player_agent_id, self.staggering_force)
-                    if has_staggering_force:
-                        yield from self.swap_to_scythe()
-                        yield from Routines.Yield.Skills.CastSkillID(self.eremites_attack, aftercast_delay=750)
-                        self.spiked = True
-                        self.spiking = False
-                        return
-
-            if self.spiked:
-                print('Cleaning up survivors, Noone lives!')
-                if target_sensali:
-                    if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.vow_of_strength)):
-                        yield from Routines.Yield.Skills.CastSkillID(self.vow_of_strength, aftercast_delay=750)
-                        return
-                    GLOBAL_CACHE.Player.Interact(target_sensali, False)
-                    yield
-                    return
-
-                remaining_enemies = Routines.Agents.GetFilteredEnemyArray(
-                    player_pos[0], player_pos[1], Range.Earshot.value
-                )
-                next_sensali = self.get_sensali_target(remaining_enemies)
-                if not target_sensali and next_sensali:
-                    if (yield from Routines.Yield.Skills.IsSkillIDUsable(self.vow_of_strength)):
-                        yield from Routines.Yield.Skills.CastSkillID(self.vow_of_strength, aftercast_delay=750)
-                        return
-                    GLOBAL_CACHE.Player.Interact(next_sensali, False)
-                    yield
-                    return
-
-                if not next_sensali:
-                    self.status = DervBuildFarmStatus.Move
-                    return
-
-
-# =================== BUILD END ========================
+from Bots.DervFeatherFarmer import DervFeatherFarmer
+from Bots.DervFeatherFarmer import DervBuildFarmStatus
 
 
 selected_step = 0
+bot_current_step = 0
 SEITUING_HARBOR = "Seitung Harbor"
 STATUS = 'status'
 SENSALI = 'Sensali'
 
 bot = Botting("Feather Farmer", custom_build=DervFeatherFarmer())
+stuck_timer = ThrottledTimer(3000)
+movement_check_timer = ThrottledTimer(3000)
+stuck_counter = 0
+old_player_position = (0, 0)
+initial_run = True
 
 
 def return_to_outpost():
@@ -272,6 +78,7 @@ def kill_sensalis(bot, kill_immediately=False):
         sensali_array = [agent_id for agent_id in enemy_array if SENSALI in GLOBAL_CACHE.Agent.GetName(agent_id)]
 
         if len(sensali_array) == 0:
+            bot.config.build_handler.__setattr__(STATUS, DervBuildFarmStatus.Move)
             break  # all sensalis dead
 
         # Timeout check
@@ -295,14 +102,78 @@ def death_or_completion_callback(bot):
     bot.States.JumpToStepName("[H]Farm Loop_2")
 
 
-def Routine(bot):
+def handle_stuck(bot):
+    global stuck_timer
+    global movement_check_timer
+    global old_player_position
+    global stuck_counter
+    global initial_run
+
+    while True:
+        # Wait until map is valid
+        if not Routines.Checks.Map.MapValid() and Routines.Checks.Map:
+            yield from Routines.Yield.wait(1000)
+            continue
+
+        if GLOBAL_CACHE.Agent.IsDead(GLOBAL_CACHE.Player.GetAgentID()):
+            return
+
+        if GLOBAL_CACHE.Map.GetMapID() == 196 and bot.config.build_handler.status == DervBuildFarmStatus.Move:
+
+            # Check if character hasn't moved
+            if movement_check_timer.IsExpired():
+                current_player_pos = GLOBAL_CACHE.Player.GetXY()
+                if old_player_position == current_player_pos:
+                    GLOBAL_CACHE.Player.SendChatCommand("stuck")
+                    print("Player is stuck!")
+                    # Pause FSM and remember where we are
+                    fsm = bot.config.FSM
+                    current_step = fsm.get_current_state_number()
+                    target_name = fsm.get_state_name_by_number(current_step)
+
+                    fsm.pause()
+
+                    print('Killing sensalis that got me stuck')
+                    # Deal with local enemies before resuming
+                    yield from kill_sensalis(bot, kill_immediately=True)
+
+                    print('Resetting to state', target_name)
+                    # Jump back into the same step
+                    fsm.jump_to_state_by_name(target_name)
+                    fsm.resume()
+                    yield
+                else:
+                    old_player_position = current_player_pos
+                    stuck_timer.Reset()
+                    stuck_counter = 0
+
+                movement_check_timer.Reset()
+
+            # Hard reset if too many consecutive stuck detections
+            if stuck_counter >= 10:
+                print("Too many stuck detections, forcing reset.")
+                stuck_counter = 0
+                fsm = bot.config.FSM
+                yield from fsm._reset_execution()
+
+        yield from Routines.Yield.wait(500)
+
+
+def load_skill_bar(bot):
+    yield from bot.config.build_handler.LoadSkillBar()
+
+
+def main_farm(bot):
     bot.Properties.Enable("auto_combat")
     bot.Events.OnDeathCallback(lambda: death_or_completion_callback(bot))
+    bot.States.AddManagedCoroutine('Stuck handler', lambda: handle_stuck(bot))
 
+    bot.States.AddHeader('Starting Loop')
     map_id = GLOBAL_CACHE.Map.GetMapID()
     if map_id != 250:
         bot.Map.Travel(target_map_name=SEITUING_HARBOR)
         bot.Wait.ForMapLoad(target_map_name=SEITUING_HARBOR)
+    bot.States.AddCustomState(lambda: load_skill_bar(bot), "Loading Skillbar")
 
     bot.Move.XY(16570, 17713, "Exit Outpost for resign spot")
     bot.Wait.ForMapLoad(target_map_name="Jaya Bluffs")
@@ -320,76 +191,81 @@ def Routine(bot):
     bot.Move.XY(1540, -6995, 'Move spot 2')
 
     bot.Move.XY(-472, -4342, 'Move to Kill Spot 1')
+    bot.States.AddHeader('Kill Spot 1')
     bot.States.AddCustomState(lambda: kill_sensalis(bot, kill_immediately=True), "Killing Sensalis Immediately")
 
     bot.Move.XY(-1536, -1686, "Move to Kill Spot 2")
+    bot.States.AddHeader('Kill Spot 2')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(586, -76, "Move to Kill Spot 3")
+    bot.States.AddHeader('Kill Spot 3')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-1556, 2786, "Move to Kill Spot 4")
+    bot.States.AddHeader('Kill Spot 4')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-2229, -815, "Move to Kill Spot 5")
+    bot.States.AddHeader('Kill Spot 5')
     bot.States.AddCustomState(lambda: kill_sensalis(bot, kill_immediately=True), "Killing Sensalis Immediately")
 
     bot.Move.XY(-5247, -3290, "Move to Kill Spot 6")
+    bot.States.AddHeader('Kill Spot 6')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-6994, -2273, "Move to Kill Spot 7")
+    bot.States.AddHeader('Kill Spot 7')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-5042, -6638, "Move to Kill Spot 8")
+    bot.States.AddHeader('Kill Spot 8')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-11040, -8577, "Move to Kill Spot 9")
+    bot.States.AddHeader('Kill Spot 9')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-10860, -2840, "Move to Kill Spot 10")
+    bot.States.AddHeader('Kill Spot 10')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-14900, -3000, "Move to Kill Spot 11")
+    bot.States.AddHeader('Kill Spot 11')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-12200, 150, "Move to Kill Spot 12")
+    bot.States.AddHeader('Kill Spot 12')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-12500, 4000, "Move to Kill Spot 13")
+    bot.States.AddHeader('Kill Spot 13')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-12111, 1690, "Move to Kill Spot 14")
+    bot.States.AddHeader('Kill Spot 14')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-10303, 4110, "Move to Kill Spot 15")
+    bot.States.AddHeader('Kill Spot 15')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-10500, 5500, "Move to Kill Spot 16")
+    bot.States.AddHeader('Kill Spot 16')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
 
     bot.Move.XY(-9700, 2400, "Move to Kill Spot 17")
+    bot.States.AddHeader('Kill Spot 17')
     bot.States.AddCustomState(lambda: kill_sensalis(bot), "Killing Sensalis")
+    bot.States.AddCustomState(lambda: death_or_completion_callback(bot), "Return to Seitung Harbor")
 
-    bot.States.AddCustomState(death_or_completion_callback, "Return to Seitung Harbor")
 
-
-bot.Routine = Routine.__get__(bot)
+bot.SetMainRoutine(main_farm)
 
 
 def main():
-    global selected_step
-
     bot.Update()
-
-    if PyImGui.begin("Feather Farmer", PyImGui.WindowFlags.AlwaysAutoResize):
-
-        if PyImGui.button("start bot"):
-            bot.Start()
-
-        if PyImGui.button("stop bot"):
-            bot.Stop()
-
-    PyImGui.end()
+    bot.UI.draw_window()
 
 
 if __name__ == "__main__":
